feat: introduce sparo-lib
