+++ apps/sparo-lib/README.md
@@ -1,0 +1,6 @@
+## sparo-lib
+
+This is a companion package for the Sparo tool.
+
+The **sparo-lib** version number is always exactly equal
+to the **sparo** version number that it depends on.
