+++ apps/sparo-lib/src/api/Sparo.ts
@@ -1,0 +1,19 @@
+import { SparoCommandLine } from '../cli/SparoCommandLine';
+import { SparoCICommandLine } from '../cli/SparoCICommandLine';
+
+/**
+ * General operations for Sparo engine.
+ *
+ * @public
+ */
+export class Sparo {
+  private constructor() {}
+
+  public static async launchSparoAsync(): Promise<void> {
+    await SparoCommandLine.launchAsync();
+  }
+
+  public static async launchSparoCIAsync(): Promise<void> {
+    await SparoCICommandLine.launchAsync();
+  }
+}
