+++ apps/sparo-lib/src/cli/SparoCICommandLine.ts
@@ -1,28 +1,30 @@
 import 'reflect-metadata';
 import { getFromContainer, registerClass } from '../di/container';
-import { GitService } from '../services/GitService';
 import { CommandService } from '../services/CommandService';
+import { CI_COMMAND_LIST } from './commands/cmd-list';
+import { ICommand } from './commands/base';
 import { ArgvService } from '../services/ArgvService';
-import { COMMAND_LIST } from '../commands/cmd-list';
-import { HelpCommand } from '../commands/help';
-import { ICommand } from '../commands/base';
+import { CIHelpCommand } from './commands/ci-help';
+import { GitVersionCompatibility } from '../logic/GitVersionCompatibility';
 
-class Sparo {
+export class SparoCICommandLine {
   private _commandsMap: Set<string> = new Set<string>();
 
   private constructor() {}
 
   public static async launchAsync(): Promise<void> {
-    const sparo: Sparo = new Sparo();
-    await sparo.prepareCommandAsync();
-    await sparo.runAsync();
+    await GitVersionCompatibility.ensureGitVersionAsync();
+
+    const sparoCI: SparoCICommandLine = new SparoCICommandLine();
+    await sparoCI.prepareCommandAsync();
+    await sparoCI.runAsync();
   }
 
   public async prepareCommandAsync(): Promise<void> {
     const commandsService: CommandService = await getFromContainer(CommandService);
 
     await Promise.all(
-      COMMAND_LIST.map(async (cmd): Promise<void> => {
+      CI_COMMAND_LIST.map(async (cmd): Promise<void> => {
         registerClass(cmd);
         const cmdInstance: ICommand<{}> = await getFromContainer(cmd);
         commandsService.register(cmdInstance);
@@ -34,19 +36,11 @@ class Sparo {
   public async runAsync(): Promise<void> {
     const argv: ArgvService = await getFromContainer(ArgvService);
     await argv.parseArgvAsync();
-    const userInputCmdName: string = argv.getUserCommand();
-    if (!userInputCmdName) {
-      const helpCommand: HelpCommand = await getFromContainer(HelpCommand);
-      return helpCommand.handler();
-    }
 
-    // proxy to gitService
-    if (!this._supportedCommand(userInputCmdName)) {
-      const gitService: GitService = await getFromContainer(GitService);
-      const args: string[] = process.argv.slice(2);
-      gitService.executeGitCommand({
-        args
-      });
+    const userInputCmdName: string = argv.getUserCommand();
+    if (!userInputCmdName || !this._supportedCommand(userInputCmdName)) {
+      const helpCommand: CIHelpCommand = await getFromContainer(CIHelpCommand);
+      return helpCommand.handler();
     }
   }
 
@@ -54,5 +48,3 @@ class Sparo {
     return this._commandsMap.has(commandName);
   }
 }
-
-export { Sparo };
