+++ apps/sparo-lib/src/index.ts
@@ -1,0 +1,1 @@
+export { Sparo } from './api/Sparo';
