+++ apps/sparo-lib/src/logic/Stopwatch.ts
