+++ apps/sparo-lib/src/schemas/sparse-profile.schema.json
