+++ apps/sparo-lib/src/services/GitService.ts
