+++ apps/sparo-lib/src/services/GitSparseCheckoutService.ts
