+++ apps/sparo-lib/src/services/HelpTextService.ts
