+++ apps/sparo-lib/src/services/LogService.ts
