+++ apps/sparo/bin/sparo-ci
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../lib-commonjs/start.js');
+require('../lib/start.js');
