+++ common/reviews/api/sparo-lib.api.md
@@ -1,0 +1,17 @@
+## API Report File for "sparo-lib"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+// @public
+export class Sparo {
+    // (undocumented)
+    static launchSparoAsync(): Promise<void>;
+    // (undocumented)
+    static launchSparoCIAsync(): Promise<void>;
+}
+
+// (No @packageDocumentation comment for this package)
+
+```
